[Fix] #16 - Timeout 설정
